Splitting bot_manager into multiple files.
