+++ src/main/python/RLBotFramework/botmanager/bot_manager.py
@@ -6,8 +6,6 @@ from datetime import datetime, timedelta
 from RLBotFramework.utils import rate_limiter
 from RLBotFramework.utils.class_importer import ExternalClassWrapper
 from RLBotFramework.utils.logging_utils import get_logger
-from RLBotFramework.utils.structures import game_data_struct as gd
-from RLBotFramework.utils.structures.bot_input_struct import PlayerInput
 from RLBotFramework.utils.structures.game_interface import GameInterface
 from RLBotFramework.utils.structures.quick_chats import send_quick_chat, register_for_quick_chat
 
@@ -166,131 +164,3 @@ class BotManager:
 
     def pull_data_from_game(self):
         raise NotImplementedError
-
-
-class BotManagerStruct(BotManager):
-    def __init__(self, terminate_request_event, termination_complete_event, bot_configuration, name, team, index,
-                 agent_class_wrapper, agent_metadata_queue, quick_chat_queue_holder):
-        """
-        See documentation on BotManager.
-        """
-        super().__init__(terminate_request_event, termination_complete_event, bot_configuration, name, team, index,
-                         agent_class_wrapper, agent_metadata_queue, quick_chat_queue_holder)
-        self.game_tick_proto = None
-
-    def prepare_for_run(self):
-        # Set up shared memory map (offset makes it so bot only writes to its own input!) and map to buffer
-        self.bot_input = PlayerInput()
-        # Set up shared memory for game data
-        self.game_tick_packet = gd.GameTickPacket()  # We want to do a deep copy for game inputs so people don't mess with em
-
-    def call_agent(self, agent, agent_class):
-        controller_input = agent.get_output_vector(self.game_tick_packet)
-        if not controller_input:
-            raise Exception('Agent "{}" did not return a player_input tuple.'.format(agent_class.__name__))
-
-        player_input = self.bot_input
-
-        # Write all player inputs
-        player_input.fThrottle = controller_input[0]
-        player_input.fSteer = controller_input[1]
-        player_input.fPitch = controller_input[2]
-        player_input.fYaw = controller_input[3]
-        player_input.fRoll = controller_input[4]
-        player_input.bJump = controller_input[5]
-        player_input.bBoost = controller_input[6]
-        player_input.bHandbrake = controller_input[7]
-        self.game_interface.update_player_input(player_input, self.index)
-
-    def get_game_time(self):
-        return self.game_tick_packet.gameInfo.TimeSeconds
-
-    def pull_data_from_game(self):
-        self.game_interface.update_live_data_packet(self.game_tick_packet)
-
-
-class BotManagerProto(BotManager):
-    def __init__(self, terminate_request_event, termination_complete_event, bot_configuration, name, team, index,
-                 agent_class_wrapper, agent_metadata_queue, quick_chat_queue_holder):
-        """
-        See documentation on BotManager.
-        """
-        super().__init__(terminate_request_event, termination_complete_event, bot_configuration, name, team, index,
-                         agent_class_wrapper, agent_metadata_queue, quick_chat_queue_holder)
-        self.game_tick_proto = None
-
-    def call_agent(self, agent, agent_class):
-        controller_state = agent.get_output_proto(self.game_tick_proto)
-        if not controller_state:
-            raise Exception('Agent "{}" did not return a controller_state.'.format(agent_class.__name__))
-
-        self.game_interface.update_controller_state(controller_state, self.index)
-
-    def get_game_time(self):
-        return self.game_tick_proto.game_info.seconds_elapsed
-
-    def pull_data_from_game(self):
-        self.game_tick_proto = self.game_interface.update_live_data_proto()
-
-    def prepare_for_run(self):
-        pass
-
-
-class BotManagerFlatbuffer(BotManager):
-    def __init__(self, terminate_request_event, termination_complete_event, bot_configuration, name, team, index,
-                 agent_class_wrapper, agent_metadata_queue, quick_chat_queue_holder):
-        """
-        See documentation on BotManager.
-        """
-        super().__init__(terminate_request_event, termination_complete_event, bot_configuration, name, team, index,
-                         agent_class_wrapper, agent_metadata_queue, quick_chat_queue_holder)
-        self.game_tick_flat = None
-
-    def call_agent(self, agent, agent_class):
-        player_input = agent.get_output_flatbuffer(self.game_tick_flat)
-        if not player_input:
-            raise Exception('Agent "{}" did not return a player input.'.format(agent_class.__name__))
-
-        self.game_interface.update_player_input_flat(player_input)
-
-    def get_game_time(self):
-        try:
-            return self.game_tick_flat.GameInfo().SecondsElapsed()
-        except AttributeError:
-            return 0.0
-
-    def pull_data_from_game(self):
-        self.game_tick_flat = self.game_interface.update_live_data_flat()
-
-    def prepare_for_run(self):
-        pass
-
-class BotManagerIndependent(BotManager):
-    def __init__(self, terminate_request_event, termination_complete_event, bot_configuration, name, team, index,
-                 agent_class_wrapper, agent_metadata_queue, quick_chat_queue_holder):
-        """
-        See documentation on BotManager.
-        """
-        super().__init__(terminate_request_event, termination_complete_event, bot_configuration, name, team, index,
-                         agent_class_wrapper, agent_metadata_queue, quick_chat_queue_holder)
-
-    def run(self):
-        # Get bot module
-        agent, agent_class_file = self.load_agent(self.agent_class_wrapper)
-        agent.run_independently(self.terminate_request_event)
-        if hasattr(agent, 'retire'):
-            agent.retire()
-        # If terminated, send callback
-        self.termination_complete_event.set()
-
-    def call_agent(self, agent, agent_class):
-        pass
-
-    def pull_data_from_game(self):
-        pass
-
-    def get_game_time(self):
-        pass
-
-    def prepare_for_run(self):
-        pass
